% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casewhen.R
\name{casewhen}
\alias{casewhen}
\title{Generate a SQL CASE WHEN statement from a mapping file}
\usage{
casewhen(inputfile)
}
\arguments{
\item{inputfile}{Mapping dataframe OR path to the mapping file}
}
\value{
A string that represents the constructed CASE statement
}
\description{
This function constructs a CASE WHEN THEN statement from a mapping file or
dataframe. It assumes that the first column of the mapping data contains the
original WHEN values, and the second column contains the THEN values (the
values to be mapped to.)
}
