% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlist.R
\name{inlist}
\alias{inlist}
\title{Generate a SQL IN statement from a mapping file}
\usage{
inlist(inputfile = NULL, header = FALSE)
}
\arguments{
\item{inputfile}{Dataframe OR path to the mapping file}

\item{header}{If reading a csv file, TRUE if the file includes a header row,
FALSE if it does not include a header row.}
}
\value{
A string that represents the constructed CASE statement
}
\description{
This function constructs an IN statement from a mapping file or
dataframe. It assumes that the first column of the data contains the
list of values to check for.
}
\examples{
 input <- Data_Frame <- data.frame(Training = c("Strength", "Stamina",
  "Other"))
 result <- inlist(inputfile = input, header = TRUE)
}
